Move functions.lua into prototypes
