+++ omnimatter_crystal/prototypes/functions.lua
